Ensure all pools pass around ERC20s instead of ETH
The bookkeeping was previously done in the fallback functions of all the
pools. This change adds in an explicit bookkeeping call that happens
before the transfer. This will consume more gas, but the previous
implementation was careful not to rely on the contract's ETH balance, so
they now should not depend solely on the ERC20 balance.